honor configuration when setting the speech rate
up to now, configuration was read, but never actually used. This begins a series of changes, perhaps across multiple fronts, to do so.
This makes `state::new` send a message on the ssip channel, with the request to set the rate to a user defined value
warning: if somehow the ssip task isn't initialized by the point we get there, the change will be lost in the channel, or may be picked up later than intended

+++ common/src/settings/speech.rs
@@ -3,7 +3,7 @@ use serde::{Deserialize, Serialize};
 #[derive(Debug, Serialize, Deserialize)]
 #[allow(clippy::module_name_repetitions)]
 pub struct SpeechSettings {
-	pub rate: i32,
+	pub rate: i8,
 }
 impl Default for SpeechSettings {
 	fn default() -> Self {
